Move de.fraunhofer.aisec.cpg.passes.order to de.fraunhofer.aisec.cpg.passes.executionConfiguration
