+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/passes/Pass.kt
@@ -35,8 +35,8 @@ import de.fraunhofer.aisec.cpg.graph.declarations.TranslationUnitDeclaration
 import de.fraunhofer.aisec.cpg.graph.scopes.Scope
 import de.fraunhofer.aisec.cpg.helpers.Benchmark
 import de.fraunhofer.aisec.cpg.helpers.SubgraphWalker.ScopedWalker
-import de.fraunhofer.aisec.cpg.passes.order.RequiredFrontend
-import de.fraunhofer.aisec.cpg.passes.order.RequiresLanguageTrait
+import de.fraunhofer.aisec.cpg.passes.executionConfiguration.RequiredFrontend
+import de.fraunhofer.aisec.cpg.passes.executionConfiguration.RequiresLanguageTrait
 import java.util.concurrent.CompletableFuture
 import java.util.function.Consumer
 import kotlin.reflect.KClass
