+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/passes/executionConfiguration/DependsOn.kt
@@ -23,19 +23,20 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.order
+package de.fraunhofer.aisec.cpg.passes.executionConfiguration
 
-import de.fraunhofer.aisec.cpg.TranslationConfiguration
 import de.fraunhofer.aisec.cpg.passes.Pass
 import kotlin.reflect.KClass
 
 /**
- * Register a new default pass required by a frontend. Passes annotated this way are collected by
- * [TranslationConfiguration.Builder.registerExtraFrontendPasses] and automatically registered in
- * [TranslationConfiguration.Builder.build], but only if
- * [TranslationConfiguration.Builder.defaultPasses] was called.
+ * Register a dependency for the annotated pass. This ensures that:
+ * - the annotated pass is executed after its dependency when `softDependency` is `false`:
+ * - the dependency is added to the list of active passes even if not manually specified by the user
+ *   when `softDependency`is true:
+ * - the dependency is not added to the list of active passes - the order is only enforced if the
+ *   user manually adds the pass
  */
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.CLASS)
 @Repeatable
-annotation class RegisterExtraPass(val value: KClass<out Pass<*>>)
+annotation class DependsOn(val value: KClass<out Pass<*>>, val softDependency: Boolean = false)
