+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/passes/executionConfiguration/PassWithDependencies.kt
@@ -23,7 +23,7 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.order
+package de.fraunhofer.aisec.cpg.passes.executionConfiguration
 
 import de.fraunhofer.aisec.cpg.graph.Node
 import de.fraunhofer.aisec.cpg.passes.Pass
