+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/passes/executionConfiguration/ReplacePass.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024, Fraunhofer AISEC. All rights reserved.
+ * Copyright (c) 2023, Fraunhofer AISEC. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,16 +23,25 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.order
+package de.fraunhofer.aisec.cpg.passes.executionConfiguration
 
 import de.fraunhofer.aisec.cpg.frontends.Language
-import de.fraunhofer.aisec.cpg.frontends.LanguageTrait
+import de.fraunhofer.aisec.cpg.passes.EvaluationOrderGraphPass
+import de.fraunhofer.aisec.cpg.passes.Pass
 import kotlin.reflect.KClass
 
 /**
- * This annotation can only enable a pass if its target language implements a given [LanguageTrait].
+ * This annotation can be used to replace a certain [Pass] (identified by [old]) for a specific
+ * [Language] (identified by [lang]) with another [Pass] (identified by [with]).
  *
- * This annotation is [Repeatable]. In this case, all specified language traits must exist on the
- * [Language].
+ * The primary use-case for this annotation is to allow language frontends to override specific
+ * passes, such as the [EvaluationOrderGraphPass] in order to optimize language specific graphs.
  */
-@Repeatable annotation class RequiresLanguageTrait(val value: KClass<out LanguageTrait>)
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.CLASS)
+@Repeatable
+annotation class ReplacePass(
+    val old: KClass<out Pass<*>>,
+    val lang: KClass<out Language<*>>,
+    val with: KClass<out Pass<*>>
+)
