+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/passes/executionConfiguration/RequiresLanguageTrait.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Fraunhofer AISEC. All rights reserved.
+ * Copyright (c) 2024, Fraunhofer AISEC. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,20 +23,16 @@
  *                    \______/ \__|       \______/
  *
  */
-package de.fraunhofer.aisec.cpg.passes.order
+package de.fraunhofer.aisec.cpg.passes.executionConfiguration
 
-import de.fraunhofer.aisec.cpg.passes.Pass
+import de.fraunhofer.aisec.cpg.frontends.Language
+import de.fraunhofer.aisec.cpg.frontends.LanguageTrait
 import kotlin.reflect.KClass
 
 /**
- * Register a dependency for the annotated pass. This ensures that:
- * - the annotated pass is executed after its dependency when `softDependency` is `false`:
- * - the dependency is added to the list of active passes even if not manually specified by the user
- *   when `softDependency`is true:
- * - the dependency is not added to the list of active passes - the order is only enforced if the
- *   user manually adds the pass
+ * This annotation can only enable a pass if its target language implements a given [LanguageTrait].
+ *
+ * This annotation is [Repeatable]. In this case, all specified language traits must exist on the
+ * [Language].
  */
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.CLASS)
-@Repeatable
-annotation class DependsOn(val value: KClass<out Pass<*>>, val softDependency: Boolean = false)
+@Repeatable annotation class RequiresLanguageTrait(val value: KClass<out LanguageTrait>)
